backend/requirements.txt
Django==4.2.16
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-gis==1.0
django-filter==24.2
django-cors-headers==4.4.0
python-decouple==3.8
